REST Mesthod "readdatapoint DONE! usage = http://localhost:8080/eu.smartcampus.api.rest/deviceapi/readdatapoint/<address> <address> = {library, kernel_14, kernel_16, room_1.17, room_1.19, UTA_A4, amphitheater_A4,  amphitheater_A5, laboratory_1.58"}

+++ eu.smartcampus.api/src/eu/smartcampus/util/Reading.java
@@ -3,15 +3,15 @@ package eu.smartcampus.util;
 
 public class Reading {
 
-    private final int value;
+    private final double value;
     private final long ts;
 
-    public Reading(int value, long ts) {
+    public Reading(double value, long ts) {
         this.value = value;
         this.ts = ts;
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
